requirements.txt
accelerate==0.20.0
adapter-transformers==3.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
awsebcli==3.20.9
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bibtexparser==1.4.0
bitsandbytes==0.39.1
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
blobfile==2.0.2
botocore==1.31.40
cachetools==5.3.1
cement==2.8.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
colorama==0.4.3
comm==0.1.3
cryptography==41.0.3
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.18.1
exceptiongroup==1.1.2
executing==1.2.0
fake-useragent==1.2.1
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
Flask-Login==0.6.2
fqdn==1.5.1
free-proxy==1.1.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
langchain==0.0.277
langsmith==0.0.30
lit==16.0.6
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
medisearch-client==0.1.9
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
ndg-httpsclient==0.5.1
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.1
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
OpenAIAuth==0.3.2
openapi-schema-pydantic==1.2.4
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
passlib==1.7.4
pathspec==0.10.1
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
pymed==0.8.9
pymongo==4.5.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
replicate==0.11.0
requests==2.28.2
revChatGPT==2.3.14
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
scholarly==1.7.11
scikit-learn==1.2.1
scipy==1.11.1
selenium==4.11.2
semantic-version==2.8.5
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.49
stack-data==0.6.2
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
termcolor==1.1.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
tiktoken==0.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
